fix lldb-dap attach test case
Summary:
The test cases added to TestDAP_attach.py in D70663297 was problematic .

The test was failed when the sandcastle task test the lldb.
The attach test fails because it tries to attach to "a.out" and there are a bunch of those running during the first test run. But could be only one on a re-run.

This fix removed the unnecessary code to launch the lldb-dap.

Test Plan:
1. lldb-dotest -f TestDAP_attach.TestDAP_attach.test_session_id_update
2. ninja check-lldb : P1785663029, this test case didn't fail.

Reviewers: peix, jalalonde, #lldb_team

Reviewed By: peix

Differential Revision: https://phabricator.intern.facebook.com/D72996648

Tasks: T221156034